Added red color on warn, moved logic to a dedicated file
