+++ CHANGELOG
@@ -1,5 +1,6 @@
 == 2.1.8 / ???
 
+* added red color for error messages.
 * added commented lines in the generated deploy.rb file to set verbosity as
   before.
 * BC break: verbose mode is disabled by default (log level = IMPORTANT), so
